fix: wrap web-sales pages as client components
