+++ app/web-sales/input/page.tsx
@@ -1,14 +1,7 @@
-"use client"
+// server wrapper for /web-sales/input
 export const dynamic = 'force-dynamic'
-import Sidebar from "@/components/sidebar"
-import WebSalesInputView from "@/components/web-sales-input-view"
-export default function WebSalesInputPage() {
-  return (
-    <div className="flex min-h-screen">
-      <Sidebar />
-      <div className="flex-1 ml-64 p-8">
-        <WebSalesInputView />
-      </div>
-    </div>
-  )
+import NextDynamic from 'next/dynamic'
+const InputClient = NextDynamic(() => import('./page.client'), { ssr: false })
+export default function Page() {
+  return <InputClient />
 }
