+++ components/WebSalesDataTable.tsx
@@ -7,7 +7,7 @@ import { WebSalesData } from "@/types/db"
 import { Plus, Trash2, TrendingUp, TrendingDown, Edit } from "lucide-react"
 import ProductAddModal from "./ProductAddModal"
 import ProductEditModal from "./ProductEditModal"
-import { supabase } from "../lib/supabase"
+import { getSupabaseBrowserClient } from "@/lib/supabase/browser"
 
 interface WebSalesDataTableProps {
  filteredItems: WebSalesData[]
@@ -52,6 +52,7 @@ export default function WebSalesDataTable({
  historicalPriceData = [],
  month,
 }: WebSalesDataTableProps) {
+ const supabase = getSupabaseBrowserClient();
  const [isAddingProduct, setIsAddingProduct] = useState(false)
  const [isEditingProduct, setIsEditingProduct] = useState(false)
  const [editingProductData, setEditingProductData] = useState<any>(null)
