+++ sales-report-form.tsx
@@ -11,7 +11,7 @@ import { Textarea } from "@/components/ui/textarea"
 import { Calendar } from "@/components/ui/calendar"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
 import { CalendarIcon, ClipboardCheck, Loader2 } from "lucide-react"
-import { supabase } from "@/lib/supabase"
+import { getSupabaseBrowserClient } from "@/lib/supabase/browser"
 import { formatDateJST, formatCurrency } from "@/lib/utils"
 
 // --- 変更点1: PropsにonSaveSuccessを追加 ---
@@ -39,6 +39,7 @@ type FormData = {
 
 // --- 変更点2: コンポーネントがonSaveSuccessを受け取るように変更 ---
 export default function SalesReportForm({ initialDate, onSaveSuccess }: SalesReportFormProps) {
+  const supabase = getSupabaseBrowserClient();
   const getInitialDate = () => {
     if (initialDate) {
       return new Date(initialDate + 'T00:00:00');
